fix: extract stuff as constants and add storybook
